Update installation.rst (#118)
Update installation from source on Windows

I propose a more detailed "Installation on MS Windows" section. 
Re-installing on my freshly formatted PC, I noticed that a few of the tricks required are not explicitly stated. An experienced user can always debug it down until it works of course, but what is the point of a documentation for the install, if not saving time in avoiding trivial mistakes? :) 

I have also made the (maybe obvious) suggestion of avoiding the install of the whole Visual Studio IDE and limit to components to save (if precious) disk space, before deciding to move to Python 2.x

* Update installation.rst

Co-authored-by: Boxi Li <etamin1201@gmail.com>

* Update installation.rst

Co-authored-by: Boxi Li <etamin1201@gmail.com>

Co-authored-by: Boxi Li <etamin1201@gmail.com>