Define appropriate LaTeX macros
For `latex*` make targets we can just load up the full `physics` package
since we're running through a full LaTeX compiler.  This isn't an option
for HTML targets, because we have to render through MathJax.  Currently
(Sphinx 3.5) the loaded version of MathJax is 2.7, which does not
support the `physics` package.  Sphinx 4, when it arrives, will load
MathJax 3, which _does_ support in.  In the meantime, we just have to
define any commands that we're going to use.