Remove idea that Jupyter/Spyder are dependencies
The previous version of the installation guide loosely suggested that
IPython, Jupyter and Spyder were optional dependencies of QuTiP.  This
is not the case, and the documentation was more meant to help people who
were coming from Matlab and new to the Python scientific ecosystem.

It had started causing some confusion with new contributors as we grew
larger, so we make it clearer these are simple suggestions.